Graph export: add labels to the export CSV format
We want to have the labels in the edge dataset to differentiate between
the file names and branch names that have the same src/dst.

This changes the format of the edge files to be:

<SRC> <DST> [LABEL] [PERMISSION]

Where LABEL is an optional base64-encoded label of the file or branch
name, and PERMISSION an integer in base 10 corresponding to the
permission of the file.