Add ORC exporter
This new exporter allows us to export the SWH graph dataset as a set of
relational tables in a static columnar format called ORC. This can then
be uploaded on data processing engines like Amazon Athena, BigQuery or
Azure Databricks.

This replaces the old scripts that were extracting the data directly
from the PostgreSQL database, to be integrated to the journal instead.
A notable change is that we now use the ORC format instead of Parquet,
as it supports streamed writes, which simplifies the data buffering and
allows for larger dataset files.