Refactor export paths in the base Exporter class
Previously, each exporter had to be given a custom __init__ argument
containing the export path. This behavior is now shared in the base
Exporter class.

This required a change in the way GraphEdgesExporter behaved, since it
is no longer instanciated with an export path that already contains the
obj_type. Its behavior is now closer to ORCExporter, where a new
writer is created on the fly in the appropriate directory for each
object type.