Encode TimestampWithTimezone as (timestamp, offset, raw_offset_bytes) in ORC file
ie. use the standard ORC Timestamp format (aka a couple
(seconds, nanoseconds)) with 2 extra fields for the offset.

The offset is stored as an integer (in minutes), but the raw offset
value is also present as a binary string representation, following
recent evolutions of swh-model.

This makes swh-dataset compatible with swh-model 5.