Add a “I will respect the ToS” button to the dataset index
As it might be easy to overlook that the dataset we provide are covered
by both an Ethical Charter and Terms of Usage for bulk access, put the
index behind a “I will respect the Ethical Charter and Terms of Use”.

This in done via some simple JavaScript that hides the index sections,
and add a button that will make them reappear on click. We remember
the click for next vists using the browser local storage.