Fetch all offsets before exporting any object
This lowers the number of orphan nodes.

Additionally, Kafka consumers occasionally hang forever on the .close() call after
getting offsets, and changing the other this way seems to fix it.

(And even if it does not fix the issue, this change makes them hang shortly after
the process started so it can be restarted quickly, instead of waiting hours before
it hangs.)