Truncate long full names instead of dropping them
Full names will be truncated instead of skipped, which will make the number of full names the same as the number of people.

Having pseudonymized authors/committers in the revision and release tables that did not have a corresponding entry in the fullname table makes it harder for swh-graph to check for inconsistencies in its inputs.