Make export_name / export_id significant
Making them insignificant means that if we abort a run and start a new one within minutes,
the Luigi worker will merge both tasks and use the oldest id.

This is undesirable because we end up with a mix of export_id within the new run,
causing an assertion failure.